refactor(devtools): migrate monolith to NuxtPage child routes, bump ^5.2.4
- split pages/link-checker.vue (inline v-if tabs) into a DevtoolsLayout shell
  + pages/link-checker/{index,links,debug,docs}.vue child routes
- lift shared link computeds (nodes/failingNodes/internalLinks/...) into state.ts
- explicit nuxtseo-layer-devtools/composables/* imports; isProductionMode redirect
- real ServerFunctions/ClientFunctions/LinkInspectionResult types from src
- tsconfig: exclude dist; bump layer -> ^5.2.4
