+++ devtools/pages/link-checker/docs.vue
@@ -1,0 +1,3 @@
+<template>
+  <DevtoolsDocs url="https://nuxtseo.com/link-checker" />
+</template>
