+++ pnpm-workspace.yaml
@@ -1,3 +1,6 @@
+minimumReleaseAgeExclude:
+  - nuxtseo-layer-devtools
+  - nuxtseo-shared
 shellEmulator: true
 
 trustPolicy: no-downgrade
@@ -27,8 +30,8 @@ catalog:
   magic-string: ^0.30.21
   nuxt: ^4.4.2
   nuxt-site-config: ^4.0.8
-  nuxtseo-layer-devtools: ^5.2.0
-  nuxtseo-shared: ^5.2.0
+  nuxtseo-layer-devtools: ^5.2.4
+  nuxtseo-shared: ^5.2.4
   ofetch: ^1.5.1
   pathe: ^2.0.3
   pkg-types: ^2.3.0
